Release 1.8 -Added Legacy4J Settings Screen, a way to search all the settings, accessible via Mods screen. -Added World Host Friends screen to Main menu. -Added `Advanced Options Mode` slider to Legacy4J Settings screen Advanced User Interface section. -Added `Save Cache` option to Game Options section. -Added `Inverted Front Camera Pitch` option to Game Options section. -Added `Auto Save Countdown` option to User Interface section. -Added `Head Follows the Camera` option to Graphics section. -Added `Fast Leaves When Blocked` option to Graphics section. -Added `Fast Leaves Custom Models` option to Graphics section. -Added `Override Terrain Fog End` option to `Graphics` section. -Added `Legacy Sky Shape` option to `Graphics` section. -Added back `Virtual Cursor` option to Controller section. -Added Legacy Common Config (`.minecraft/config/legacy/common.json`). -Added Legacy Common Mixin Config (`.minecraft/config/legacy/common_mixin.json`). -Added Legacy Mixin Options (`.minecraft/config/legacy/client_mixin.json`). -Added Legacy4J Settings key mapping, with `Y` as default key. -Added `Display System Messages as Overlay` option in Advanced User Interface section. -Added `Global Resource Packs` to Graphics section. -Added `Classic Trading`, `Classic Stonecutting`, `Classic Loom` and `Mixed Crafting` options to User Interface section. -Added missing `Force Smooth Movement` Portuguese translations. -Added `put_scrollable_renderer`, `put_renderable_vertical_list`, `modify_renderable_vertical_list`, `draw_outlined_string` and `render_enchanted_book` UI Definition Element Types. -Added `Random Block Rotations` and `Nearest Mipmap Scaling` Factory API options to Graphics section. -Added manual mipmaps for water. -Added a Patch Notes screen, which appears when the game/Legacy4J is updated or launched for the first time. -Added common options to Game Options section (or Host Options when playing in a world): Legacy Combat`, `Legacy Sword Blocking` and `Squared View Distance`. -Added client mixin options to Graphics section:  -Screens: `Legacy Inventory Screen`, `Legacy Classic Crafting Screen`,  `Legacy Classic Stonecutter Screen`, `Legacy Classic Loom Screen`, `Legacy Classic Merchant Screen`, `Legacy Container Screen`, `Legacy Crafter Screen` (in versions >=1.20.4), `Legacy Furnace Screen`, `Legacy Anvil Screen`, `Legacy Smithing Screen`, `Legacy Gridstone Screen`, `Legacy Cartography Screen`, `Legacy Enchantment Screen`, `Legacy Beacon Screen`, `Legacy Brewing Stand Screen`, `Legacy Book Screen`, `Legacy Create World Screen` and `Legacy Title Screen`.  -Miscellaneous: `Legacy Gui`, `Legacy Chat`, `Legacy Bosshealth`, `Legacy Witches` and `Legacy Drowned`. -Now, the leaves when using Fast graphics will use an alternative model instead of removing the cutout effect, similar to LE. -Now, Legacy Waters has a new logo. -Now, the Mixed Crafting Interface will be used when Legacy4J isn't present on server side. -Now, the Alphabetical sorting will be used in the Mods screen by default. -Now, the Graphics section will show the resource selectors tooltip box. -Now, the `Interface Resolution` slider will be nominally 50 to 150%, equivalent to the real value used. -Now, the enchantment tooltips will be similar to LE. -Now, the options are completely unlinked from vanilla, having a new directory `.minecraft/config/legacy/client_options.json`. -Now, the How To Play will be similar to LE (thanks to Moustard for providing the text) and almost completely data-driven with UI Definitions. -Now, it'll be possible to access the vanilla Options screen via Mods screen. -Now, the sugar cane won't have tint when using Programmer Art or Console Aspects, being similar to LE. -Now, multiline edit boxes will have a LE-style and allow using the keyboard. -Now, the Add Album screen will allow description changes. -Now, any resolution next to 720p will use the mojangles 11 font. -Now, the Brewing Stand Interface panel will be positioned similar to LE. -Now, the Controller Mapping and Keybinds screens are more similar to LE, having tooltips for cancel and conflicts. -Now, the Loading and Load Save screen are more similar to LE. -Now, the text in tabs is positioned more similarly to LE. -Now, there will be a message when trying to spawn the wither in peaceful mode, like in LE. -Now, the loading background is a texture `legacy:textures/gui/loading_background.png`. -Now, hints and autosave icon will no longer appear when the gui is hidden (with F1) -Now, the `Terrain Fog End` option will be in chunks, being more similar to LE. -Improved JEI controller compatibility. -Renamed the Resource Assorts to Resource Albums (thanks to cloud54 for suggesting this). -Removed blockstate definitions that disabled the random block rotations. -Removed all the deprecated listing syntax. -Fixed incompatibility with Hold My Items mod. -Fixed Controller Manager stopping working when using more than one controller on Linux and macOS. -Fixed a lot of confirmation screens having incorrect background panel height. -Fixed Player Tab List with wrong background height. -Fixed Programmer Art and Console Aspects not having item definitions in 1.21.2+, resulting in broken models. -Fixed Interface Opacity being different in versions 1.21.2+. -Fixed crash when loading a World Template with a non-game path. -Fixed Loading Overlay screen not rendering correctly in versions +1.21.2. -Fixed Firework with incorrect texture format in versions +1.21.2. -Fixed mixin crash in versions forge/neoforge <=1.20.4. -Fixed items with custom names based on the item stack not having a default tip based on their name, such as potions. -Fixed `put_legacy_slot` UI Element Type missing the warning icon when it's a real slot, and not updating its display when it's a fake slot. -Fixed Merchant Interface with weird scale when the villager can't level up. -Fixed scroll out of the panel in the Host Options Interface. -Fixed chat line spacing not being at max by default in versions >=1.21.2. -Fixed Creative Mode screen not allowing to drop items in versions >=1.21.2. -Fixed JEI having wrong positions for recipe viewers. -Fixed loading background not being black in versions >=1.21.4. -Fixed incompatibility with Iris shaders caused by `Flying View Rolling`. -Fixed a lot of mixin incompatibilities by using mixin extras. -Fixed items and blocks being invisible in the hotbar when using Fabulous graphics in 1.21.1. -Updated Mod Credits. -MIT License is being used again.

+++ src/main/java/wily/legacy/mixin/base/client/create_world/CreateWorldScreenMixin.java
@@ -1,5 +1,6 @@
 package wily.legacy.mixin.base.client.create_world;
 
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
@@ -23,6 +24,8 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+import wily.factoryapi.base.ArbitrarySupplier;
+import wily.factoryapi.base.Bearer;
 import wily.factoryapi.base.client.UIAccessor;
 import wily.legacy.Legacy4JClient;
 import wily.legacy.client.CommonColor;
@@ -49,7 +52,7 @@ public abstract class CreateWorldScreenMixin extends Screen implements ControlTo
 
     @Shadow protected abstract void onCreate();
 
-    protected boolean trustPlayers = true;
+    protected Bearer<Boolean> trustPlayers = Bearer.of(true);
     protected Panel panel;
     protected PublishScreen publishScreen;
     protected PackAlbum.Selector resourceAssortSelector;
@@ -72,7 +75,7 @@ public abstract class CreateWorldScreenMixin extends Screen implements ControlTo
     @Override
     public void added() {
         super.added();
-        resourceAssortSelector.addControlTooltips(this,ControlTooltip.Renderer.of(this));
+        OptionsScreen.setupSelectorControlTooltips(ControlTooltip.Renderer.of(this), this);
     }
 
     @Inject(method = "init", at = @At("HEAD"), cancellable = true)
@@ -94,7 +97,7 @@ public abstract class CreateWorldScreenMixin extends Screen implements ControlTo
             difficultyButton.active = !uiState.isHardcore();
         });
 
-        addRenderableWidget(Button.builder(Component.translatable( "createWorld.tab.more.title"), button -> minecraft.setScreen(new WorldMoreOptionsScreen(self(), b-> trustPlayers = b))).bounds(panel.x + 13, panel.y + 172,220,20).build());
+        addRenderableWidget(Button.builder(Component.translatable( "createWorld.tab.more.title"), button -> minecraft.setScreen(new WorldMoreOptionsScreen(self(), trustPlayers))).bounds(panel.x + 13, panel.y + 172,220,20).build());
         addRenderableWidget(Button.builder(Component.translatable("selectWorld.create"), button -> this.onCreate()).bounds(panel.x + 13, panel.y + 197,220,20).build());
         addRenderableWidget(new TickBox(panel.x+ 14, panel.y+155,220,publishScreen.publish, b-> PublishScreen.PUBLISH, b->null, button -> {
             if (button.selected) minecraft.setScreen(publishScreen);
@@ -106,30 +109,33 @@ public abstract class CreateWorldScreenMixin extends Screen implements ControlTo
         this.uiState.onChanged();
     }
 
-    @Override
-    public void repositionElements() {
+    @Inject(method = "repositionElements", at = @At("HEAD"), cancellable = true)
+    public void repositionElements(CallbackInfo ci) {
         rebuildWidgets();
+        ci.cancel();
     }
 
     @Inject(method = /*? if >=1.21.2 {*/ /*"createWorldAndCleanup"*//*?} else {*/"createNewWorld"/*?}*/,at = @At("RETURN"))
     private void onCreate(CallbackInfo ci) {
+        resourceAssortSelector.applyChanges(true);
         Legacy4JClient.serverPlayerJoinConsumer = s->{
-            LegacyClientWorldSettings.of(s.server.getWorldData()).setTrustPlayers(trustPlayers);
+            LegacyClientWorldSettings.of(s.server.getWorldData()).setTrustPlayers(trustPlayers.get());
             s.server.getPlayerList().sendPlayerPermissionLevel(s);
             publishScreen.publish((IntegratedServer) s.server);
             LegacyClientWorldSettings.of(minecraft.getSingleplayerServer().getWorldData()).setSelectedResourceAlbum(resourceAssortSelector.getSelectedAlbum());
         };
     }
 
-    @Redirect(method = "createNewWorld", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/Minecraft;createWorldOpenFlows()Lnet/minecraft/client/gui/screens/worldselection/WorldOpenFlows;"))
-    private WorldOpenFlows createNewWorld(Minecraft instance) {
-        return new WorldOpenFlows(minecraft,Legacy4JClient.getLevelStorageSource());
+    @ModifyExpressionValue(method = "createNewWorld", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/Minecraft;createWorldOpenFlows()Lnet/minecraft/client/gui/screens/worldselection/WorldOpenFlows;"))
+    private WorldOpenFlows createNewWorld(WorldOpenFlows original) {
+        return LegacyOptions.saveCache.get() ? new WorldOpenFlows(minecraft,Legacy4JClient.currentWorldSource) : original;
     }
 
     @Inject(method = "createNewWorldDirectory", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/storage/LevelStorageSource;createAccess(Ljava/lang/String;)Lnet/minecraft/world/level/storage/LevelStorageSource$LevelStorageAccess;"))
     private /*? >=1.21.2 {*//*static*//*?}*/ void createNewWorldDirectory(/*? >=1.21.2 {*//*Minecraft minecraft, String string, @Nullable Path path, *//*?}*/CallbackInfoReturnable<Optional<LevelStorageSource.LevelStorageAccess>> cir) {
+        if (!LegacyOptions.saveCache.get()) return;
         try {
-            LevelStorageSource.LevelStorageAccess access = Legacy4JClient.getLevelStorageSource().createAccess(/*? <1.21.2 {*/uiState.getTargetFolder()/*?} else {*//*string*//*?}*/);
+            LevelStorageSource.LevelStorageAccess access = Legacy4JClient.currentWorldSource.createAccess(/*? <1.21.2 {*/uiState.getTargetFolder()/*?} else {*//*string*//*?}*/);
             access.close();
             if (Files.exists(access.getDimensionPath(Level.OVERWORLD))) FileUtils.deleteDirectory(access.getDimensionPath(Level.OVERWORLD).toFile());
         } catch (IOException e) {
@@ -137,9 +143,9 @@ public abstract class CreateWorldScreenMixin extends Screen implements ControlTo
         }
     }
 
-    @Redirect(method = "createNewWorldDirectory", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/Minecraft;getLevelSource()Lnet/minecraft/world/level/storage/LevelStorageSource;"))
-    private /*? if >=1.21.2 {*//*static*//*?}*/ LevelStorageSource createNewWorldDirectory(Minecraft instance) {
-        return Legacy4JClient.currentWorldSource;
+    @ModifyExpressionValue(method = "createNewWorldDirectory", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/Minecraft;getLevelSource()Lnet/minecraft/world/level/storage/LevelStorageSource;"))
+    private /*? if >=1.21.2 {*//*static*//*?}*/ LevelStorageSource createNewWorldDirectory(LevelStorageSource original) {
+        return LegacyOptions.saveCache.get() ? Legacy4JClient.currentWorldSource : original;
     }
 
     //? if >1.20.1 {
